This extra test breaks the build after #6792 has been merged (#6816)
* This extra test breaks the build after #6792 has been merged: the fix that that issue breaks on 2 filesystems, under conditions similar to the original problem but other filesystems than LocalFileSystem.

* Fix FileSystemActor and unit tests

* 6792 review comments

---------

+++ filesystem/src/test/java/org/frankframework/filesystem/LocalFileSystemTestHelper.java
@@ -14,14 +14,15 @@ public class LocalFileSystemTestHelper implements IFileSystemTestHelper {
 
 	public Path folder;
 
-
 	public LocalFileSystemTestHelper(Path folder2) {
 		this.folder=folder2;
 	}
 
 	@Override
 	public void setUp() throws Exception {
-		FileUtils.cleanDirectory(folder.toFile());
+		if (Files.exists(folder)) {
+			FileUtils.cleanDirectory(folder.toFile());
+		}
 	}
 
 	@Override
